dom/messagechannel/tests/test_messageChannel_pref.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=677638
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 677638 - pref</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=677638">Mozilla Bug 677638</a>
<div id="content"></div>
<pre id="test">
</pre>
  <script type="application/javascript">

  function runTest(what) {
    var status;
    try {
      status = MessageChannel;
      ok(what, "Should MessageChannel exist?");
    } catch(e) {
      ok(!what, "Should MessageChannel exist?");
    }
  }

  SimpleTest.waitForExplicitFinish();

  runTest(false);
  SpecialPowers.pushPrefEnv({"set": [["dom.messageChannel.enabled", true]]},
                            function() { runTest(true); SimpleTest.finish(); });
  </script>
</body>
</html>
